backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.9
pytest==8.3.3
httpx==0.27.2
# Pin bcrypt to stable version compatible with Python 3.8+
bcrypt==4.0.1
argon2-cffi==23.1.0
# Rate limiting
slowapi==0.1.9
# Email support
aiosmtplib==3.0.1
